requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
cryptography==41.0.7
django-filter==23.4
drf-yasg==1.21.7
faker==20.1.0
factory-boy==3.2.1
pytest==7.4.3
pytest-env==1.1.3
pytest-django==4.7.0
gunicorn==21.2.0
